d11 done

+++ Advent of Code 2023/main.cpp
@@ -7,7 +7,7 @@ int main()
 {
 	Solutions* AOC = new Solutions;
 	vector<string> lines = AOC->getInput("11");
-	int part1 = 0, part2 = 0;
+	s64 part1 = 0, part2 = 0;
 	
 	constexpr int loops = 1;
 	auto p1_t1 = chrono::high_resolution_clock::now();
